Moved models as suggested by Enyst
