+++ openhands/storage/conversation/file_conversation_store.py
@@ -5,12 +5,12 @@ from dataclasses import dataclass
 
 from openhands.core.config.app_config import AppConfig
 from openhands.core.logger import openhands_logger as logger
-from openhands.server.data_models.conversation_metadata_result_set import (
+from openhands.storage.data_models.conversation_metadata_result_set import (
     ConversationMetadataResultSet,
 )
 from openhands.storage import get_file_store
 from openhands.storage.conversation.conversation_store import ConversationStore
-from openhands.server.data_models.conversation_metadata import ConversationMetadata
+from openhands.storage.data_models.conversation_metadata import ConversationMetadata
 from openhands.storage.files import FileStore
 from openhands.storage.locations import (
     CONVERSATION_BASE_DIR,
