+++ tests/unit/test_conversation.py
@@ -5,11 +5,11 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from openhands.server.data_models.conversation_info import ConversationInfo
-from openhands.server.data_models.conversation_info_result_set import (
+from openhands.storage.data_models.conversation_info import ConversationInfo
+from openhands.storage.data_models.conversation_info_result_set import (
     ConversationInfoResultSet,
 )
-from openhands.server.data_models.conversation_status import ConversationStatus
+from openhands.storage.data_models.conversation_status import ConversationStatus
 from openhands.server.routes.new_conversation import (
     get_conversation,
     search_conversations,
